Redesign the `resumable` class
- Introduce event IDs for `resume`: this addition allows an resumable to
  support multiple event types. The only use case we currently have is
  to have an explicit dispose event that we can use during shutdown. The
  scheduler API has been prepared to support user-defined event IDs but
  the default scheduler will ignore this parameter for now (and always
  pass the default event ID). We also introduce an initialization event
  ID, but this ID matters only when using the deterministic test
  fixture.
- Remove the max-throughput parameter from `resume`: this is an
  implementation detail of scheduled actors and should not leak into the
  scheduler/resumable API.
- Remove `awaiting_message`: returning `awaiting_message` from `resume`
  does exactly the same as returning `done`.
- Remove `shutdown_execution_unit`: abusing the return type of `resume`
  for shutdown logic is hacky and pollutes the interface. Using an
  explicit shutdown flag internally is much cleaner and does not leak
  internal implementation details into the public API.
- Remove `subtype`: the only use case we had for this tag was to make
  sure that brokers only run on the multiplexer and not on the system
  scheduler and that regular actors do not run on a multiplexer.
  However, pinning a resumable (broker) to their dedicated scheduler
  (multiplexer) is much more general and robust.

+++ libcaf_core/caf/async/spsc_buffer.hpp
@@ -448,7 +448,7 @@ public:
     return owner_ ? action::state::scheduled : action::state::disposed;
   }
 
-  resume_result resume(scheduler*, size_t) override {
+  resume_result resume(scheduler*, uint64_t) override {
     on_wakeup_t on_wakeup;
     {
       std::unique_lock guard{mtx_};
@@ -722,7 +722,7 @@ public:
     return owner_ ? action::state::scheduled : action::state::disposed;
   }
 
-  resume_result resume(scheduler*, size_t) override {
+  resume_result resume(scheduler*, uint64_t) override {
     on_demand_t on_demand;
     size_t demand = 0;
     buffer_ptr_t buf;
